rearrange event route and splitted into several components
