+++ web/src/routes/Events/components/index.jsx
@@ -1,0 +1,2 @@
+export { default as TableHead } from './tableHead';
+export { default as Filters } from './filters';
